requirements.txt
anyio==3.7.0
asyncio==3.4.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
fastapi==0.99.1
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
mysql==0.0.3
mysqlclient==2.2.0
oauthlib==3.2.2
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.10
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
waitress==2.1.2
